requirements.txt
pyodbc==5.1.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
openai==1.54.0
chromadb==0.5.18
sentence-transformers==3.3.1
beautifulsoup4==4.12.3
python-telegram-bot==21.6
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.1
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.8
tiktoken==0.8.0
